refactor(app): move member addition to workspace repository
