+++ crates/coral-app/src/state/db/workspace_member_state.rs
@@ -2,7 +2,6 @@
 
 #![cfg_attr(not(test), expect(dead_code, reason = "used higher in the PR stack"))]
 
-use super::repositories::users::UserRecord;
 use super::{CoralDb, DbError, DbRepos};
 use crate::workspaces::MemberRole;
 
@@ -31,59 +30,6 @@ pub(crate) enum RemoveMemberOutcome {
 }
 
 impl CoralDb {
-    pub(crate) async fn add_workspace_member(
-        &self,
-        workspace_id: &str,
-        user_id: &str,
-        role: MemberRole,
-        created_at_unix_nanos: i64,
-    ) -> Result<AddMemberOutcome, DbError> {
-        let mut tx = self.begin().await?;
-        if !tx
-            .workspaces()
-            .hold_for_child_mutation(workspace_id)
-            .await?
-        {
-            tx.rollback().await?;
-            return Ok(AddMemberOutcome::WorkspaceNotFound);
-        }
-        let Some(user) = tx.users().get_by_user_id(user_id).await? else {
-            tx.rollback().await?;
-            return Ok(AddMemberOutcome::UserNotFound);
-        };
-        if let Some(existing_role) = tx
-            .workspace_members()
-            .role_for_user_id(workspace_id, user_id)
-            .await?
-        {
-            tx.rollback().await?;
-            return Ok(existing_add_outcome(user, existing_role, role));
-        }
-        match tx
-            .workspace_members()
-            .insert(workspace_id, user_id, role, created_at_unix_nanos)
-            .await
-        {
-            Ok(()) => {
-                tx.commit().await?;
-                Ok(AddMemberOutcome::Added(member_view(user, role)))
-            }
-            Err(error) if error.is_unique_violation() => {
-                tx.rollback().await?;
-                let mut session = self;
-                let Some(existing_role) = session
-                    .workspace_members()
-                    .role_for_user_id(workspace_id, user_id)
-                    .await?
-                else {
-                    return Err(error);
-                };
-                Ok(existing_add_outcome(user, existing_role, role))
-            }
-            Err(error) => Err(error),
-        }
-    }
-
     pub(crate) async fn remove_workspace_member(
         &self,
         workspace_id: &str,
@@ -119,26 +65,6 @@ impl CoralDb {
             tx.rollback().await?;
             Ok(RemoveMemberOutcome::MemberNotFound)
         }
-    }
-}
-
-fn existing_add_outcome(
-    user: UserRecord,
-    existing_role: MemberRole,
-    requested_role: MemberRole,
-) -> AddMemberOutcome {
-    if existing_role == requested_role {
-        AddMemberOutcome::ExistingSameRole(member_view(user, existing_role))
-    } else {
-        AddMemberOutcome::RoleConflict
-    }
-}
-
-fn member_view(user: UserRecord, role: MemberRole) -> WorkspaceMemberView {
-    WorkspaceMemberView {
-        user_id: user.user_id,
-        role,
-        display_name: user.display_name,
     }
 }
 
@@ -195,14 +121,18 @@ mod tests {
             WorkspaceCreationOutcome::Created
         );
 
+        let mut first_session = db;
+        let mut second_session = db;
+        let mut first_workspaces = first_session.workspaces();
+        let mut second_workspaces = second_session.workspaces();
         let (first, second) = tokio::join!(
-            db.add_workspace_member(
+            first_workspaces.add_member(
                 &identical_add_workspace_id,
                 &member_id,
                 MemberRole::Member,
                 11,
             ),
-            db.add_workspace_member(
+            second_workspaces.add_member(
                 &identical_add_workspace_id,
                 &member_id,
                 MemberRole::Member,
@@ -232,14 +162,18 @@ mod tests {
                 .expect("create conflicting-add workspace"),
             WorkspaceCreationOutcome::Created
         );
+        let mut member_session = db;
+        let mut owner_session = db;
+        let mut member_workspaces = member_session.workspaces();
+        let mut owner_workspaces = owner_session.workspaces();
         let (member_add, owner_add) = tokio::join!(
-            db.add_workspace_member(
+            member_workspaces.add_member(
                 &conflicting_add_workspace_id,
                 &conflicting_member_id,
                 MemberRole::Member,
                 21,
             ),
-            db.add_workspace_member(
+            owner_workspaces.add_member(
                 &conflicting_add_workspace_id,
                 &conflicting_member_id,
                 MemberRole::Owner,
@@ -292,15 +226,18 @@ mod tests {
                 .expect("create owner-removal workspace"),
             WorkspaceCreationOutcome::Created
         );
+        let mut session = db;
         assert!(matches!(
-            db.add_workspace_member(
-                &owner_removal_workspace_id,
-                &other_owner_id,
-                MemberRole::Owner,
-                31,
-            )
-            .await
-            .expect("add owner"),
+            session
+                .workspaces()
+                .add_member(
+                    &owner_removal_workspace_id,
+                    &other_owner_id,
+                    MemberRole::Owner,
+                    31,
+                )
+                .await
+                .expect("add owner"),
             AddMemberOutcome::Added(_)
         ));
